Fix issue with loading Prisma user in the Edge
