+++ auth.config.ts
@@ -1,10 +1,10 @@
 import type { NextAuthConfig } from "next-auth";
 import Google from "next-auth/providers/google";
 import GitHub from "next-auth/providers/github";
-import Credentials from "next-auth/providers/credentials";
-import { compare } from "bcryptjs";
-import prisma from "@/prisma/prismaClient";
 
+// Edge-compatible configuration
+// This file is imported by middleware which runs in Edge Runtime
+// DO NOT import Prisma or any Node.js-only dependencies here
 export default {
   trustHost: true,
   pages: {
@@ -20,31 +20,6 @@ export default {
   providers: [
     Google,
     GitHub,
-    Credentials({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Password", type: "password" },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) return null;
-
-        const user = await prisma.user.findUnique({ where: { email: credentials.email as string } });
-
-        if (!user || !user.password) return null;
-
-        const isValid = await compare(credentials.password as string, user.password);
-
-        if (!isValid) return null;
-
-        return {
-          id: user.id,
-          name: user.name,
-          email: user.email,
-          image: user.image,
-        };
-      },
-    }),
   ],
 } satisfies NextAuthConfig;
